Hack MSVC static library paths to avoid clashes
The default waf configuration uses .lib for both static libraries and the
manifest for DLLs, which is (somehow) the convention on Windows, but causes
build errors because multiple targets produce the same outputs.

Switch to using the UNIX convention ".a" (which is not unheard of on Windows,
and is the approach used by Meson) to avoid this problem.